reorganized all the routes into multiple files
