+++ server/requirements.txt
@@ -3,3 +3,4 @@ Flask-SQLAlchemy==3.1.1
 Werkzeug==3.0.3
 pyotp==2.9.0
 mysqlclient==2.2.4
+Flask-SocketIO==5.3.6
